@@ -136,7 +136,7 @@ namespace Microsoft.TemplateEngine.Cli.HelpAndUsage
                 return null;
             }
 
-            TemplateListResolver.ParseTemplateArgs(templateInfo, hostDataLoader, commandInput);
+            TemplateResolver.ParseTemplateArgs(templateInfo, hostDataLoader, commandInput);
             allParams = templateCreator.SetupDefaultParamValuesFromTemplateAndHost(template, template.DefaultName ?? "testName", out IReadOnlyList<string> defaultParamsWithInvalidValues);
             templateCreator.ResolveUserParameters(template, allParams, commandInput.InputTemplateParams, out userParamsWithInvalidValues);
             hasPostActionScriptRunner = CheckIfTemplateHasScriptRunningPostActions(template, environmentSettings, commandInput, templateCreator);
@@ -151,7 +151,7 @@ namespace Microsoft.TemplateEngine.Cli.HelpAndUsage
                 {
                     commandInput.InputTemplateParams.TryGetValue(canonical, out string specifiedValue);
                     string inputFormat = commandInput.TemplateParamInputFormat(canonical);
-                    InvalidParameterInfo invalidParam = new InvalidParameterInfo(inputFormat, specifiedValue, canonical);
+                    InvalidParameterInfo invalidParam = new InvalidParameterInfo(InvalidParameterInfo.Kind.InvalidParameterValue, inputFormat, specifiedValue, canonical);
                     invalidParameters.Add(invalidParam);
                 }
             }
@@ -172,7 +172,7 @@ namespace Microsoft.TemplateEngine.Cli.HelpAndUsage
                             ?? inputVariants.Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur)
                             ?? param.Name;
 
-                        InvalidParameterInfo invalidParam = new InvalidParameterInfo(displayName, param.DefaultValue, displayName, true);
+                        InvalidParameterInfo invalidParam = new InvalidParameterInfo(InvalidParameterInfo.Kind.InvalidDefaultValue, displayName, param.DefaultValue, displayName);
                         invalidParameters.Add(invalidParam);
                     }
                 }
