@@ -10,20 +10,17 @@ namespace Microsoft.TemplateEngine.Cli.TemplateSearch
 {
     public class CliHostSpecificDataMatchFilterFactory
     {
-        public CliHostSpecificDataMatchFilterFactory(INewCommandInput commandInput, string defaultLanguage)
+        public CliHostSpecificDataMatchFilterFactory(INewCommandInput commandInput)
         {
             _commandInput = commandInput;
-            _defaultLanguage = defaultLanguage;
         }
 
         private readonly INewCommandInput _commandInput;
-        private readonly string _defaultLanguage;
 
-        public Func<IReadOnlyList<ITemplateNameSearchResult>, IReadOnlyList<ITemplateMatchInfo>> MatchFilter => (nameMatches) =>
+        public Func<IReadOnlyList<ITemplateNameSearchResult>, IReadOnlyList<ITemplateMatchInfo>> MatchFilter => (foundPackages) =>
         {
             Dictionary<string, HostSpecificTemplateData> hostDataLookup = new Dictionary<string, HostSpecificTemplateData>();
-
-            foreach (ITemplateNameSearchResult result in nameMatches)
+            foreach (ITemplateNameSearchResult result in foundPackages)
             {
                 if (result is CliTemplateNameSearchResult cliResult)
                 {
@@ -34,19 +31,16 @@ namespace Microsoft.TemplateEngine.Cli.TemplateSearch
                     hostDataLookup[result.Template.Identity] = HostSpecificTemplateData.Default;
                 }
             }
-
             IHostSpecificDataLoader hostSpecificDataLoader = new InMemoryHostSpecificDataLoader(hostDataLookup);
 
-            TemplateListResolutionResult templateResolutionResult = TemplateListResolver.GetTemplateResolutionResult(nameMatches.Select(x => x.Template).ToList(), hostSpecificDataLoader, _commandInput, _defaultLanguage);
+            IEnumerable<Func<INewCommandInput, Func<PackInfo, bool>>> packageFiltersToUse = SupportedFilterOptions.SupportedSearchFilters
+                                    .OfType<PackageFilterOption>()
+                                    .Select(filter => filter.PackageMatchFilter);
 
-            if (templateResolutionResult.TryGetAllInvokableTemplates(out IReadOnlyList<ITemplateMatchInfo> invokableTemplates))
-            {
-                return invokableTemplates;
-            }
-            else
-            {
-                return new List<ITemplateMatchInfo>();
-            }
+            IEnumerable<ITemplateNameSearchResult> templatesToFilter =
+                foundPackages.Where(foundPackage => packageFiltersToUse.All(pf => pf(_commandInput)(foundPackage.PackInfo)));
+
+            return TemplateResolver.PerformCoreTemplateQueryForSearch(templatesToFilter.Select(x => x.Template), hostSpecificDataLoader, _commandInput).ToList();
         };
     }
 }
