@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-#if !NET45
+#if !NETFULL
 using System.Runtime.Loader;
 #endif
 
@@ -10,7 +10,7 @@ namespace Microsoft.TemplateEngine.Edge
 {
     public static class AssemblyLoadContextHelper
     {
-#if !NET45
+#if !NETFULL
         public static IEnumerable<KeyValuePair<string, Assembly>> LoadAllFromCodebase(this AssemblyLoadContext context, Paths paths, out IEnumerable<string> loadFailures, string pattern = "*.dll", SearchOption searchOption = SearchOption.AllDirectories)
 #else
         public static IEnumerable<KeyValuePair<string, Assembly>> LoadAllFromCodebase(this AppDomain context, Paths paths, out IEnumerable<string> loadFailures, string pattern = "*.dll", SearchOption searchOption = SearchOption.AllDirectories)
@@ -19,7 +19,7 @@ namespace Microsoft.TemplateEngine.Edge
             return LoadAllFromPath(context, paths, out loadFailures, paths.Global.BaseDir, pattern, searchOption);
         }
 
-#if !NET45
+#if !NETFULL
         public static IEnumerable<KeyValuePair<string, Assembly>> LoadAllFromPath(this AssemblyLoadContext context, Paths paths, out IEnumerable<string> loadFailures, string path, string pattern = "*.dll", SearchOption searchOption = SearchOption.AllDirectories)
 #else
         public static IEnumerable<KeyValuePair<string, Assembly>> LoadAllFromPath(this AppDomain context, Paths paths, out IEnumerable<string> loadFailures, string path, string pattern = "*.dll", SearchOption searchOption = SearchOption.AllDirectories)
@@ -34,7 +34,7 @@ namespace Microsoft.TemplateEngine.Edge
                 {
                     Assembly assembly = null;
 
-#if !NET45
+#if !NETFULL
                     if(file.IndexOf("netcoreapp", StringComparison.OrdinalIgnoreCase) > -1 || file.IndexOf("netstandard", StringComparison.OrdinalIgnoreCase) > -1)
                     {
                         using (Stream fileStream = paths.OpenRead(file))
