@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net.Http;
+using System.Threading.Tasks;
 using Microsoft.TemplateSearch.TemplateDiscovery.PackProviders;
 using Newtonsoft.Json;
 
@@ -9,7 +10,7 @@ namespace Microsoft.TemplateSearch.TemplateDiscovery.Nuget
 {
     public class NugetPackProvider : IPackProvider
     {
-        private static readonly string SearchUrlFormat = "https://api-v2v3search-0.nuget.org/query?q=template&skip={0}&take={1}&prerelease={2}";
+        private string _searchUriFormat;
         // {PackageId}.{Version}.nupkg
         private static readonly string DownloadUrlFormat = "https://api.nuget.org/v3-flatcontainer/{0}/{1}/{0}.{1}.nupkg";
         private static readonly string DownloadPackageFileNameFormat = "{0}.{1}.nupkg";
@@ -17,16 +18,18 @@ namespace Microsoft.TemplateSearch.TemplateDiscovery.Nuget
         private readonly string _packageTempPath;
         private readonly int _pageSize;
         private readonly bool _runOnlyOnePage;
-        private readonly bool _includePreviewPacks;
 
         private static readonly string DownloadedPacksDir = "DownloadedPacks";
 
-        public NugetPackProvider(string packageTempBasePath, int pageSize, bool runOnlyOnePage, bool includePreviewPacks)
+        public string Name { get; private set; }
+
+        public NugetPackProvider(string name, string query, string packageTempBasePath, int pageSize, bool runOnlyOnePage, bool includePreviewPacks)
         {
+            Name = name;
             _pageSize = pageSize;
             _runOnlyOnePage = runOnlyOnePage;
-            _packageTempPath = Path.Combine(packageTempBasePath, DownloadedPacksDir);
-            _includePreviewPacks = includePreviewPacks;
+            _packageTempPath = Path.Combine(packageTempBasePath, DownloadedPacksDir, Name);
+            _searchUriFormat = $"https://api-v2v3search-0.nuget.org/query?{query}&skip={{0}}&take={{1}}&prerelease={includePreviewPacks}";
 
             if (Directory.Exists(_packageTempPath))
             {
@@ -38,53 +41,32 @@ namespace Microsoft.TemplateSearch.TemplateDiscovery.Nuget
             }
         }
 
-        public IEnumerable<IInstalledPackInfo> CandidatePacks
+        public async IAsyncEnumerable<IPackInfo> GetCandidatePacksAsync()
         {
-            get
+            int skip = 0;
+            bool done = false;
+            int packCount = 0;
+
+            do
             {
-                int skip = 0;
-                bool done = false;
-                int packCount = 0;
-
-                do
+                string queryString = string.Format(_searchUriFormat, skip, _pageSize);
+                Uri queryUri = new Uri(queryString);
+                using (HttpClient client = new HttpClient())
+                using (HttpResponseMessage response = await client.GetAsync(queryUri).ConfigureAwait(false))
                 {
-                    string queryString = string.Format(SearchUrlFormat, skip, _pageSize, _includePreviewPacks);
-                    Uri queryUri = new Uri(queryString);
-                    using (HttpClient client = new HttpClient())
-                    using (HttpResponseMessage response = client.GetAsync(queryUri).Result)
+                    if (response.IsSuccessStatusCode)
                     {
-                        if (response.IsSuccessStatusCode)
+                        string responseText = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+
+                        NugetPackageSearchResult resultsForPage = JsonConvert.DeserializeObject<NugetPackageSearchResult>(responseText);
+
+                        if (resultsForPage.Data.Count > 0)
                         {
-                            string responseText = response.Content.ReadAsStringAsync().Result;
-
-                            NugetPackageSearchResult resultsForPage = JsonConvert.DeserializeObject<NugetPackageSearchResult>(responseText);
-
-                            if (resultsForPage.Data.Count > 0)
+                            skip += _pageSize;
+                            packCount += resultsForPage.Data.Count;
+                            foreach (NugetPackageSourceInfo sourceInfo in resultsForPage.Data)
                             {
-                                skip += _pageSize;
-                                packCount += resultsForPage.Data.Count;
-
-                                foreach (NugetPackageSourceInfo sourceInfo in resultsForPage.Data)
-                                {
-                                    if (TryDownloadPackage(sourceInfo, out string packageFilePath))
-                                    {
-                                        NugetPackInfo packInfo = new NugetPackInfo()
-                                        {
-                                            VersionedPackageIdentity = sourceInfo.VersionedPackageIdentity,
-                                            Id = sourceInfo.Id,
-                                            Version = sourceInfo.Version,
-                                            Path = packageFilePath,
-                                            TotalDownloads = sourceInfo.TotalDownloads
-                                            
-                                        };
-
-                                        yield return packInfo;
-                                    }
-                                }
-                            }
-                            else
-                            {
-                                done = true;
+                                yield return sourceInfo;
                             }
                         }
                         else
@@ -92,28 +74,16 @@ namespace Microsoft.TemplateSearch.TemplateDiscovery.Nuget
                             done = true;
                         }
                     }
-                } while (!done && !_runOnlyOnePage);
-            }
-        }
-
-        public int CandidatePacksCount
-        {
-            get
-            {
-                string queryString = string.Format(SearchUrlFormat, 0, _pageSize, _includePreviewPacks);
-                Uri queryUri = new Uri(queryString);
-                using (HttpClient client = new HttpClient())
-                using (HttpResponseMessage response = client.GetAsync(queryUri).Result)
-                {
-                    response.EnsureSuccessStatusCode();
-                    string responseText = response.Content.ReadAsStringAsync().Result;
-                    NugetPackageSearchResult resultsForPage = JsonConvert.DeserializeObject<NugetPackageSearchResult>(responseText);
-                    return resultsForPage.TotalHits;
+                    else
+                    {
+                        done = true;
+                    }
                 }
-            }
+            } while (!done && !_runOnlyOnePage);
+
         }
 
-        private bool TryDownloadPackage(NugetPackageSourceInfo packinfo, out string packageFilePath)
+        public async Task<IDownloadedPackInfo> DownloadPackageAsync(IPackInfo packinfo)
         {
             string downloadUrl = string.Format(DownloadUrlFormat, packinfo.Id, packinfo.Version);
             string packageFileName = string.Format(DownloadPackageFileNameFormat, packinfo.Id, packinfo.Version);
@@ -123,18 +93,37 @@ namespace Microsoft.TemplateSearch.TemplateDiscovery.Nuget
             {
                 using (HttpClient client = new HttpClient())
                 {
-                    byte[] packageBytes = client.GetByteArrayAsync(downloadUrl).Result;
+                    byte[] packageBytes = await client.GetByteArrayAsync(downloadUrl).ConfigureAwait(false);
                     File.WriteAllBytes(outputPackageFileNameFullPath, packageBytes);
+                    return new NugetPackInfo()
+                    {
+                        Id = packinfo.Id,
+                        Version = packinfo.Version,
+                        Path = outputPackageFileNameFullPath,
+                        TotalDownloads = packinfo.TotalDownloads
+
+                    };
                 }
             }
-            catch
+            catch (Exception e)
             {
-                packageFilePath = null;
-                return false;
+                Console.WriteLine($"Failed to download package {packinfo.Id} {packinfo.Version}, reason: {e.ToString()}.");
+                return null;
             }
+        }
 
-            packageFilePath = outputPackageFileNameFullPath;
-            return true;
+        public async Task<int> GetPackageCountAsync()
+        {
+            string queryString = string.Format(_searchUriFormat, 0, _pageSize);
+            Uri queryUri = new Uri(queryString);
+            using (HttpClient client = new HttpClient())
+            using (HttpResponseMessage response = await client.GetAsync(queryUri).ConfigureAwait(false))
+            {
+                response.EnsureSuccessStatusCode();
+                string responseText = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+                NugetPackageSearchResult resultsForPage = JsonConvert.DeserializeObject<NugetPackageSearchResult>(responseText);
+                return resultsForPage.TotalHits;
+            }
         }
 
         public void DeleteDownloadedPacks()
