@@ -12,6 +12,7 @@ using Microsoft.TemplateEngine.Cli.UnitTests.CliMocks;
 using System.Linq;
 using Microsoft.TemplateEngine.Orchestrator.RunnableProjects;
 using Microsoft.TemplateEngine.Cli.HelpAndUsage;
+using Microsoft.TemplateEngine.Mocks;
 
 namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
 {
@@ -21,30 +22,12 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_UniqueNameMatchesCorrectly()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console1",
-                Name = "Long name for Console App",
-                Identity = "Console.App",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>()
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console2",
-                Name = "Long name for Console App #2",
-                Identity = "Console.App2",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>()
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console1", name: "Long name for Console App", identity: "Console.App"));
+            templatesToSearch.Add(new MockTemplateInfo("console2", name: "Long name for Console App #2", identity: "Console.App2"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console2",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console2").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal("console2", matchResult.UnambiguousTemplateGroup.Single().Info.ShortName);
@@ -56,30 +39,12 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_ExactMatchOnShortNameMatchesCorrectly()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>()
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console2",
-                Name = "Long name for Console App #2",
-                Identity = "Console.App2",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>()
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App"));
+            templatesToSearch.Add(new MockTemplateInfo("console2", name: "Long name for Console App #2", identity: "Console.App2"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal("console", matchResult.UnambiguousTemplateGroup.Single().Info.ShortName);
@@ -91,52 +56,15 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_UnambiguousGroupIsFound()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L3",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L2", groupIdentity: "Console.App.Test").WithTag("language", "L2"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L3", groupIdentity: "Console.App.Test").WithTag("language", "L3"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(1, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(3, matchResult.ExactMatchedTemplates.Count);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal(3, matchResult.UnambiguousTemplateGroup.Count);
@@ -146,76 +74,17 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_MultipleGroupsAreFound()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L3",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "classlib",
-                Name = "Long name for Class Library App",
-                Identity = "Class.Library.L1",
-                GroupIdentity = "Class.Library.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "classlib",
-                Name = "Long name for Class Library App",
-                Identity = "Class.Library.L2",
-                GroupIdentity = "Class.Library.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L2", groupIdentity: "Console.App.Test").WithTag("language", "L2"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L3", groupIdentity: "Console.App.Test").WithTag("language", "L3"));
+            templatesToSearch.Add(new MockTemplateInfo("classlib", name: "Long name for Class Library App", identity: "Class.Library.L1", groupIdentity: "Class.Library.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("classlib", name: "Long name for Class Library App", identity: "Class.Library.L2", groupIdentity: "Class.Library.Test").WithTag("language", "L2"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "c",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("c").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(2, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(2, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(5, matchResult.ExactMatchedTemplates.Count);
             Assert.False(matchResult.HasUnambiguousTemplateGroup);
         }
@@ -224,38 +93,12 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_DefaultLanguageDisambiguates()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L2", groupIdentity: "Console.App.Test").WithTag("language", "L2"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, "L1");
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, "L1");
             Assert.True(matchResult.HasExactMatches);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.True(matchResult.HasUnambiguousTemplateGroupForDefaultLanguage);
@@ -270,39 +113,12 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_InputLanguageIsPreferredOverDefault()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L2", groupIdentity: "Console.App.Test").WithTag("language", "L2"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                Language = "L2"
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console", "L2").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, "L1");
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, "L1");
             Assert.True(matchResult.HasExactMatches);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal("console", matchResult.UnambiguousTemplateGroup.Single().Info.ShortName);
@@ -315,52 +131,15 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_UnambiguousGroup_TemplatesAreNotSameLanguage()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.L3",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L2", groupIdentity: "Console.App.Test").WithTag("language", "L2"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.L3", groupIdentity: "Console.App.Test").WithTag("language", "L3"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(1, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(3, matchResult.ExactMatchedTemplates.Count);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal(3, matchResult.UnambiguousTemplateGroup.Count);
@@ -371,52 +150,15 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_UnambiguousGroup_TemplatesAreSameLanguage()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T3",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") }
-                }
-            });
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T2", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
+            templatesToSearch.Add(new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T3", groupIdentity: "Console.App.Test").WithTag("language", "L1"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(1, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(3, matchResult.ExactMatchedTemplates.Count);
             Assert.True(matchResult.HasUnambiguousTemplateGroup);
             Assert.Equal(3, matchResult.UnambiguousTemplateGroup.Count);
@@ -427,38 +169,20 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_PartialMatch_HasLanguageMismatch()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                Language = "L2"
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console", "L2").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(1, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(1, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.True(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
             Assert.False(matchResult.HasBaselineMismatch);
@@ -470,38 +194,21 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_PartialMatch_HasContextMismatch()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                TypeFilter = "item"
-            };
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            INewCommandInput userInputs = new MockNewCommandInput("console", type: "item").WithHelpOption();
+
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(1, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(1, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.True (matchResult.HasContextMismatch);
             Assert.False(matchResult.HasBaselineMismatch);
@@ -513,38 +220,20 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_PartialMatch_HasBaselineMismatch()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                BaselineName = "core"
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("console").WithHelpOption().WithCommandOption("--baseline", "core");
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(1, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(1, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
             Assert.True(matchResult.HasBaselineMismatch);
@@ -556,40 +245,21 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_PartialMatch_HasMultipleMismatches()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                Language = "L2",
-                TypeFilter = "item",
-                BaselineName = "core"
-            };
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            INewCommandInput userInputs = new MockNewCommandInput("console", "L2", "item").WithHelpOption().WithCommandOption("--baseline", "core");
+
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(1, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(1, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.True(matchResult.HasLanguageMismatch);
             Assert.True(matchResult.HasContextMismatch);
             Assert.True(matchResult.HasBaselineMismatch);
@@ -601,58 +271,27 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_PartialMatchGroup_HasTypeMismatch_HasGroupLanguageMatch()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T2",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "console",
-                IsHelpFlagSpecified = true,
-                Language = "L2",
-                TypeFilter = "item"
-            };
+            templatesToSearch.Add(
+               new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T2", groupIdentity: "Console.App.Test")
+                   .WithTag("language", "L2")
+                   .WithTag("type", "project")
+                   .WithBaselineInfo("app", "standard"));
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            INewCommandInput userInputs = new MockNewCommandInput("console", "L2", "item").WithHelpOption();
+
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(2, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(1, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.True(matchResult.HasContextMismatch);
             Assert.False(matchResult.HasBaselineMismatch);
@@ -664,42 +303,22 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_NoMatch()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
 
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "zzzzz",
-                IsHelpFlagSpecified = true,
-                Language = "L1",
-                TypeFilter = "project",
-                BaselineName = "app"
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("zzzzz", "L1", "item").WithHelpOption().WithCommandOption("--baseline", "app");
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
             Assert.False(matchResult.HasPartialMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.Equal(0, matchResult.PartiallyMatchedTemplates.Count);
-            Assert.Equal(0, matchResult.PartiallyMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.PartiallyMatchedTemplateGroups.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
             Assert.False(matchResult.HasBaselineMismatch);
@@ -712,86 +331,32 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_OtherParameterMatch_Text()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test1",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "langVersion", new CacheParameter("text", "", "test description") }
-                }
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test1")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithParameters("langVersion")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T2",
-                GroupIdentity = "Console.App.Test2",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "test", new CacheParameter("text", "", "test description") }
-                }
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T2", groupIdentity: "Console.App.Test2")
+                    .WithTag("language", "L2")
+                    .WithTag("type", "project")
+                    .WithParameters("test")
+                    .WithBaselineInfo("app", "standard"));
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T3", groupIdentity: "Console.App.Test3")
+                    .WithTag("language", "L3")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T3",
-                GroupIdentity = "Console.App.Test3",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>()
-            });
+            INewCommandInput userInputs = new MockNewCommandInput("c").WithHelpOption().WithTemplateOption("langVersion");
 
-
-            INewCommandInput userInputs = new MockNewCommandInput(
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "langVersion", null }
-                }
-            )
-            {
-                TemplateName = "c",
-                IsHelpFlagSpecified = true,
-            };
-
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(1, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(1, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(1, matchResult.ExactMatchedTemplates.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
@@ -807,84 +372,32 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_OtherParameterMatch_Choice()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test1",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")},
-                    { "framework", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "netcoreapp1.0", "netcoreapp1.1" }) }
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>()
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test1")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithTag("framework", "netcoreapp1.0", "netcoreapp1.1")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T2",
-                GroupIdentity = "Console.App.Test2",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "test", new CacheParameter("text", "", "test description") }
-                }
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T2", groupIdentity: "Console.App.Test2")
+                    .WithTag("language", "L2")
+                    .WithTag("type", "project")
+                    .WithParameters("test")
+                    .WithBaselineInfo("app", "standard"));
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T3", groupIdentity: "Console.App.Test3")
+                    .WithTag("language", "L3")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T3",
-                GroupIdentity = "Console.App.Test3",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>()
-            });
+            INewCommandInput userInputs = new MockNewCommandInput("c").WithHelpOption().WithTemplateOption("framework");
 
-
-            INewCommandInput userInputs = new MockNewCommandInput(
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "framework", null }
-                }
-            )
-            {
-                TemplateName = "c",
-                IsHelpFlagSpecified = true,
-            };
-
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(3, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(3, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(3, matchResult.ExactMatchedTemplates.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
@@ -899,84 +412,32 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_OtherParameterDoesNotExist()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test1",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")},
-                    { "framework", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "netcoreapp1.0", "netcoreapp1.1" }) }
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>()
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test1")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithTag("framework", "netcoreapp1.0", "netcoreapp1.1")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T2",
-                GroupIdentity = "Console.App.Test2",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L2") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "test", new CacheParameter("text", "", "test description") }
-                }
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T2", groupIdentity: "Console.App.Test2")
+                    .WithTag("language", "L2")
+                    .WithTag("type", "project")
+                    .WithParameters("test")
+                    .WithBaselineInfo("app", "standard"));
 
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T3", groupIdentity: "Console.App.Test3")
+                    .WithTag("language", "L3")
+                    .WithTag("type", "project")
+                    .WithBaselineInfo("app", "standard"));
 
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T3",
-                GroupIdentity = "Console.App.Test3",
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L3") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                },
-                CacheParameters = new Dictionary<string, ICacheParameter>()
-            });
+            INewCommandInput userInputs = new MockNewCommandInput("c").WithHelpOption().WithTemplateOption("do-not-exist");
 
-
-            INewCommandInput userInputs = new MockNewCommandInput(
-                new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "do-not-exist", null }
-                }
-            )
-            {
-                TemplateName = "c",
-                IsHelpFlagSpecified = true,
-            };
-
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.True(matchResult.HasExactMatches);
-            Assert.Equal(3, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(3, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(3, matchResult.ExactMatchedTemplates.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
@@ -991,37 +452,20 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         public void TestGetTemplateResolutionResult_MatchByTagsIgnoredForHelp()
         {
             List<ITemplateInfo> templatesToSearch = new List<ITemplateInfo>();
-            templatesToSearch.Add(new TemplateInfo()
-            {
-                ShortName = "console",
-                Name = "Long name for Console App",
-                Identity = "Console.App.T1",
-                GroupIdentity = "Console.App.Test",
-                CacheParameters = new Dictionary<string, ICacheParameter>(),
-                Classifications = new List<string> { "Common", "Test" },
-                Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                {
-                    { "language", ResolutionTestHelper.CreateTestCacheTag("L1") },
-                    { "type", ResolutionTestHelper.CreateTestCacheTag("project")}
-                },
-                BaselineInfo = new Dictionary<string, IBaselineInfo>()
-                {
-                    { "app", new BaselineInfo() },
-                    { "standard", new BaselineInfo() }
-                }
-            });
+            templatesToSearch.Add(
+                new MockTemplateInfo("console", name: "Long name for Console App", identity: "Console.App.T1", groupIdentity: "Console.App.Test1")
+                    .WithTag("language", "L1")
+                    .WithTag("type", "project")
+                    .WithClassifications("Common", "Test")
+                    .WithBaselineInfo("app", "standard"));
 
 
-            INewCommandInput userInputs = new MockNewCommandInput()
-            {
-                TemplateName = "Common",
-                IsHelpFlagSpecified = true,
-            };
+            INewCommandInput userInputs = new MockNewCommandInput("Common").WithHelpOption();
 
-            ListOrHelpTemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
+            TemplateListResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResultForListOrHelp(templatesToSearch, new MockHostSpecificDataLoader(), userInputs, null);
             Assert.False(matchResult.HasExactMatches);
             Assert.False(matchResult.HasPartialMatches);
-            Assert.Equal(0, matchResult.ExactMatchedTemplatesGrouped.Count);
+            Assert.Equal(0, matchResult.ExactMatchedTemplateGroups.Count);
             Assert.Equal(0, matchResult.ExactMatchedTemplates.Count);
             Assert.False(matchResult.HasLanguageMismatch);
             Assert.False(matchResult.HasContextMismatch);
