@@ -1,12 +1,12 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Cli.CommandParsing;
 using Microsoft.TemplateEngine.Cli.TemplateResolution;
 using Microsoft.TemplateEngine.Cli.UnitTests.CliMocks;
-using Microsoft.TemplateEngine.Edge.Settings;
 using Microsoft.TemplateEngine.Edge.Template;
-using Microsoft.TemplateEngine.Orchestrator.RunnableProjects;
+using Microsoft.TemplateEngine.Mocks;
 using Xunit;
 
 namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
@@ -23,16 +23,14 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
 
             foreach (string testShortName in shortNamesForGroup)
             {
-                INewCommandInput userInputs = new MockNewCommandInput()
-                {
-                    TemplateName = testShortName
-                };
+                INewCommandInput userInputs = new MockNewCommandInput(testShortName);
 
-                TemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
-                matchResult.TryGetCoreMatchedTemplatesWithDisposition(x => x.IsMatch, out IReadOnlyList<ITemplateMatchInfo> matchedTemplateList);
-                Assert.Equal(3, matchedTemplateList.Count);
+                TemplateResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
+                Assert.Equal(TemplateResolutionResult.UnambiguousTemplateGroupStatus.SingleMatch, matchResult.GroupResolutionStatus);
+                Assert.Equal(3, matchResult.UnambiguousTemplateGroup.Templates.Count);
+                Assert.True(matchResult.UnambiguousTemplateGroup.Templates.All(t => t.IsMatch));
 
-                foreach (ITemplateMatchInfo templateMatchInfo in matchedTemplateList)
+                foreach (ITemplateMatchInfo templateMatchInfo in matchResult.UnambiguousTemplateGroup.Templates)
                 {
                     Assert.Equal("MultiName.Test", templateMatchInfo.Info.GroupIdentity);
                     Assert.Equal(1, templateMatchInfo.MatchDisposition.Count);
@@ -51,15 +49,11 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
 
             foreach (string testShortName in shortNamesForGroup)
             {
-                INewCommandInput userInputs = new MockNewCommandInput()
-                {
-                    TemplateName = testShortName
-                };
+                INewCommandInput userInputs = new MockNewCommandInput(testShortName);
 
-                TemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
-                Assert.True(matchResult.TryGetSingularInvokableMatch(out ITemplateMatchInfo invokableTemplate, out TemplateListResolutionResult.SingularInvokableMatchCheckStatus resultStatus));
-                Assert.Equal(TemplateListResolutionResult.SingularInvokableMatchCheckStatus.SingleMatch, resultStatus);
-                Assert.Equal("MultiName.Test.High.CSharp", invokableTemplate.Info.Identity);
+                TemplateResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
+                Assert.Equal(TemplateResolutionResult.Status.SingleMatch, matchResult.ResolutionStatus);
+                Assert.Equal("MultiName.Test.High.CSharp", matchResult.TemplateToInvoke.Info.Identity);
             }
         }
 
@@ -73,16 +67,11 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
 
             foreach (string testShortName in shortNamesForGroup)
             {
-                INewCommandInput userInputs = new MockNewCommandInput()
-                {
-                    TemplateName = testShortName,
-                    Language = "F#"
-                };
+                INewCommandInput userInputs = new MockNewCommandInput(testShortName, "F#");
 
-                TemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
-                Assert.True(matchResult.TryGetSingularInvokableMatch(out ITemplateMatchInfo invokableTemplate, out TemplateListResolutionResult.SingularInvokableMatchCheckStatus resultStatus));
-                Assert.Equal(TemplateListResolutionResult.SingularInvokableMatchCheckStatus.SingleMatch, resultStatus);
-                Assert.Equal("Multiname.Test.Only.FSharp", invokableTemplate.Info.Identity);
+                TemplateResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), userInputs, "C#");
+                Assert.Equal(TemplateResolutionResult.Status.SingleMatch, matchResult.ResolutionStatus);
+                Assert.Equal("Multiname.Test.Only.FSharp", matchResult.TemplateToInvoke.Info.Identity);
             }
         }
 
@@ -95,21 +84,11 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         [InlineData("eee", "Y", "Multiname.Test.Only.FSharp")]  // uses a short name from a different template in the group
         public void ChoiceValueDisambiguatesMatchesWithMultipleShortNames(string name, string fooChoice, string expectedIdentity)
         {
-            IReadOnlyDictionary<string, string> userParams = new Dictionary<string, string>()
-            {
-                {  "foo", fooChoice }
-            };
+            INewCommandInput commandInput = new MockNewCommandInput(name).WithTemplateOption("foo", fooChoice);
 
-            INewCommandInput commandInput = new MockNewCommandInput(userParams)
-            {
-                TemplateName = name
-            };
-
-            TemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), commandInput, "C#");
-
-            Assert.True(matchResult.TryGetSingularInvokableMatch(out ITemplateMatchInfo invokableTemplate, out TemplateListResolutionResult.SingularInvokableMatchCheckStatus resultStatus));
-            Assert.Equal(TemplateListResolutionResult.SingularInvokableMatchCheckStatus.SingleMatch, resultStatus);
-            Assert.Equal(expectedIdentity, invokableTemplate.Info.Identity);
+            TemplateResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), commandInput, "C#");
+            Assert.Equal(TemplateResolutionResult.Status.SingleMatch, matchResult.ResolutionStatus);
+            Assert.Equal(expectedIdentity, matchResult.TemplateToInvoke.Info.Identity);
         }
 
         [Theory(DisplayName = nameof(ParameterExistenceDisambiguatesMatchesWithMultipleShortNames))]
@@ -121,21 +100,11 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
         [InlineData("eee", "OnlyF", "someValue", "Multiname.Test.Only.FSharp")] // uses a short name from a different template in the group
         public void ParameterExistenceDisambiguatesMatchesWithMultipleShortNames(string name, string paramName, string paramValue, string expectedIdentity)
         {
-            IReadOnlyDictionary<string, string> userParams = new Dictionary<string, string>()
-            {
-                { paramName, paramValue }
-            };
+            INewCommandInput commandInput = new MockNewCommandInput(name).WithTemplateOption(paramName, paramValue);
 
-            INewCommandInput commandInput = new MockNewCommandInput(userParams)
-            {
-                TemplateName = name
-            };
-
-            TemplateListResolutionResult matchResult = TemplateListResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), commandInput, "C#");
-
-            Assert.True(matchResult.TryGetSingularInvokableMatch(out ITemplateMatchInfo invokableTemplate, out TemplateListResolutionResult.SingularInvokableMatchCheckStatus resultStatus));
-            Assert.Equal(TemplateListResolutionResult.SingularInvokableMatchCheckStatus.SingleMatch, resultStatus);
-            Assert.Equal(expectedIdentity, invokableTemplate.Info.Identity);
+            TemplateResolutionResult matchResult = TemplateResolver.GetTemplateResolutionResult(MultiShortNameGroupTemplateInfo, new MockHostSpecificDataLoader(), commandInput, "C#");
+            Assert.Equal(TemplateResolutionResult.Status.SingleMatch, matchResult.ResolutionStatus);
+            Assert.Equal(expectedIdentity, matchResult.TemplateToInvoke.Info.Identity);
         }
 
         private static IReadOnlyList<ITemplateInfo> MultiShortNameGroupTemplateInfo
@@ -146,74 +115,28 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.TemplateResolutionTests
                 {
                     List<ITemplateInfo> templateList = new List<ITemplateInfo>();
 
-                    templateList.Add(new TemplateInfo()
-                    {
-                        ShortNameList = new string[] { "aaa", "bbb" },
-                        Name = "High precedence C# in group",
-                        Precedence = 2000,
-                        Identity = "MultiName.Test.High.CSharp",
-                        GroupIdentity = "MultiName.Test",
-                        Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "language", ResolutionTestHelper.CreateTestCacheTag("C#") },
-                            { "foo", ResolutionTestHelper.CreateTestCacheTag(new string[] { "A", "W" }) }
-                        },
-                        CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "HighC", new CacheParameter("string", "high c", "high c description") }
-                        }
-                    });
+                    templateList.Add(
+                        new MockTemplateInfo(new string[] { "aaa", "bbb" }, name: "High precedence C# in group", precedence: 2000, identity: "MultiName.Test.High.CSharp", groupIdentity: "MultiName.Test")
+                            .WithTag("language", "C#")
+                            .WithTag("foo", "A", "W")
+                            .WithParameters("HighC"));
 
-                    templateList.Add(new TemplateInfo()
-                    {
-                        ShortNameList = new string[] { "ccc", "ddd", "eee" },
-                        Name = "Low precedence C# in group",
-                        Precedence = 100,
-                        Identity = "MultiName.Test.Low.CSharp",
-                        GroupIdentity = "MultiName.Test",
-                        Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "language", ResolutionTestHelper.CreateTestCacheTag("C#") },
-                            { "foo", ResolutionTestHelper.CreateTestCacheTag(new string[] { "A", "X" }) }
-                        },
-                        CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "LowC", new CacheParameter("string", "low c", "low c description") }
-                        }
-                    });
+                    templateList.Add(
+                        new MockTemplateInfo(new string[] { "ccc", "ddd", "eee" }, name: "Low precedence C# in group", precedence: 100, identity: "MultiName.Test.Low.CSharp", groupIdentity: "MultiName.Test")
+                            .WithTag("language", "C#")
+                            .WithTag("foo", "A", "X")
+                            .WithParameters("LowC"));
 
-                    templateList.Add(new TemplateInfo()
-                    {
-                        ShortNameList = new string[] { "fff" },
-                        Name = "Only F# in group",
-                        Precedence = 100,
-                        Identity = "Multiname.Test.Only.FSharp",
-                        GroupIdentity = "MultiName.Test",
-                        Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "language", ResolutionTestHelper.CreateTestCacheTag("F#") },
-                            { "foo", ResolutionTestHelper.CreateTestCacheTag(new string[] { "A", "Y" }) }
-                        },
-                        CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "OnlyF", new CacheParameter("string", "only f", "only f description") }
-                        }
-                    });
+                    templateList.Add(
+                       new MockTemplateInfo(new string[] { "fff" }, name: "Only F# in group", precedence: 100, identity: "Multiname.Test.Only.FSharp", groupIdentity: "MultiName.Test")
+                           .WithTag("language", "F#")
+                           .WithTag("foo", "A", "Y")
+                           .WithParameters("OnlyF"));
 
-                    templateList.Add(new TemplateInfo()
-                    {
-                        ShortNameList = new string[] { "other" },
-                        Name = "Unrelated template",
-                        Precedence = 9999,
-                        Identity = "Unrelated.Template.CSharp",
-                        GroupIdentity = "Unrelated.Template",
-                        Tags = new Dictionary<string, ICacheTag>(StringComparer.OrdinalIgnoreCase)
-                        {
-                            { "language", ResolutionTestHelper.CreateTestCacheTag("C#") },
-                            { "foo", ResolutionTestHelper.CreateTestCacheTag(new string[] { "A", "Z" }) }
-                        },
-                        CacheParameters = new Dictionary<string, ICacheParameter>(StringComparer.OrdinalIgnoreCase)
-                    });
+                    templateList.Add(
+                        new MockTemplateInfo(new string[] { "other" }, name: "Unrelated template", precedence: 9999, identity: "Unrelated.Template.CSharp", groupIdentity: "Unrelated.Template")
+                            .WithTag("language", "C#")
+                            .WithTag("foo", "A", "Z"));
 
                     _multiShortNameGroupTemplateInfo = templateList;
                 }
