@@ -26,10 +26,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
             EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
 
-            INewCommandInput commandInput = new MockNewCommandInput()
-            {
-                TemplateName = "foo"
-            };
+            INewCommandInput commandInput = new MockNewCommandInput("foo");
 
             TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
             SearchResults searchResults = await searchCoordinator.SearchAsync();
@@ -51,15 +48,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
             EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
 
-            // The template symbol is "Framework" (capital "F"). This checks that the host specific override is applied
-            Dictionary<string, string> rawCommandInputs = new Dictionary<string, string>()
-            {
-                { "framework", "netcoreapp2.0" }
-            };
-            INewCommandInput commandInput = new MockNewCommandInput(rawCommandInputs)
-            {
-                TemplateName = "foo"
-            };
+            INewCommandInput commandInput = new MockNewCommandInput("foo").WithTemplateOption("framework", "netcoreapp2.0");
 
             TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
             SearchResults searchResults = await searchCoordinator.SearchAsync();
@@ -69,15 +58,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             Assert.Equal(1, searchResults.MatchesBySource[0].PacksWithMatches.Count);
             Assert.Single(searchResults.MatchesBySource[0].PacksWithMatches[_packTwoInfo].TemplateMatches.Where(t => string.Equals(t.Info.Name, _fooTwoTemplate.Name)));
 
-            // same check, except with the short version of Framework, namely "f"
-            Dictionary<string, string> shortNameCheckRawCommandInputs = new Dictionary<string, string>()
-            {
-                { "f", "netcoreapp2.0" }
-            };
-            INewCommandInput shortNameCommandInput = new MockNewCommandInput(shortNameCheckRawCommandInputs)
-            {
-                TemplateName = "foo"
-            };
+            INewCommandInput shortNameCommandInput = new MockNewCommandInput("foo").WithTemplateOption("f", "netcoreapp2.0");
 
             TemplateSearchCoordinator shortNameSearchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, shortNameCommandInput, DefaultLanguage);
             SearchResults shortNameSearchResults = await searchCoordinator.SearchAsync();
@@ -97,14 +78,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
 
             // "tfm" is not a vaild symbol for the "foo" template. So it should not match.
-            Dictionary<string, string> rawCommandInputs = new Dictionary<string, string>()
-            {
-                { "tfm", "netcoreapp2.0" }
-            };
-            INewCommandInput commandInput = new MockNewCommandInput(rawCommandInputs)
-            {
-                TemplateName = "foo"
-            };
+            INewCommandInput commandInput = new MockNewCommandInput("foo").WithTemplateOption("tfm", "netcoreapp2.0");
 
             TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
             SearchResults searchResults = await searchCoordinator.SearchAsync();
@@ -122,11 +96,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
 
             Dictionary<string, string> rawCommandInputs = new Dictionary<string, string>();
-            MockNewCommandInput commandInput = new MockNewCommandInput(rawCommandInputs)
-            {
-                Language = "F#",
-                TemplateName = "bar"
-            };
+            MockNewCommandInput commandInput = new MockNewCommandInput("bar", "F#");
 
             TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
             SearchResults searchResults = await searchCoordinator.SearchAsync();
@@ -137,6 +107,89 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             Assert.Single(searchResults.MatchesBySource[0].PacksWithMatches[_packThreeInfo].TemplateMatches.Where(t => string.Equals(t.Info.Name, _barFSharpTemplate.Name)));
         }
 
+        [Theory(DisplayName = nameof(CacheSearchAuthorFilterTest))]
+        [InlineData("", "test", 1)]
+        [InlineData("foo", "test", 1)]
+        [InlineData("", "Wrong", 0)]
+        public async Task CacheSearchAuthorFilterTest(string commandTemplate, string commandAuthor, int matchCount)
+        {
+            TemplateDiscoveryMetadata mockTemplateDiscoveryMetadata = SetupDiscoveryMetadata(false);
+            MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
+            EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
+
+            MockNewCommandInput commandInput = new MockNewCommandInput(commandTemplate).WithCommandOption("--author", commandAuthor);
+
+            TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
+            SearchResults searchResults = await searchCoordinator.SearchAsync();
+
+            Assert.True(searchResults.AnySources);
+            if (matchCount == 0)
+            {
+                Assert.Equal(0, searchResults.MatchesBySource.Count);
+            }
+            else
+            {
+                Assert.Equal(1, searchResults.MatchesBySource.Count);
+                Assert.Equal(matchCount, searchResults.MatchesBySource[0].PacksWithMatches.Count);
+            }
+        }
+
+        [Theory(DisplayName = nameof(CacheSearchTypeFilterTest))]
+        [InlineData("", "project", 1)]
+        [InlineData("foo", "project", 1)]
+        [InlineData("", "Wrong", 0)]
+        public async Task CacheSearchTypeFilterTest(string commandTemplate, string commandType, int matchCount)
+        {
+            TemplateDiscoveryMetadata mockTemplateDiscoveryMetadata = SetupDiscoveryMetadata(false);
+            MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
+            EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
+
+            MockNewCommandInput commandInput = new MockNewCommandInput(commandTemplate, type: commandType);
+
+            TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
+            SearchResults searchResults = await searchCoordinator.SearchAsync();
+
+            Assert.True(searchResults.AnySources);
+            if (matchCount == 0)
+            {
+                Assert.Equal(0, searchResults.MatchesBySource.Count);
+            }
+            else
+            {
+                Assert.Equal(1, searchResults.MatchesBySource.Count);
+                Assert.Equal(matchCount, searchResults.MatchesBySource[0].PacksWithMatches.Count);
+            }
+        }
+
+        [Theory(DisplayName = nameof(CacheSearchPackageFilterTest))]
+        [InlineData("", "Three", 1, 2)]
+        [InlineData("barC", "Three", 1, 1)]
+        [InlineData("foo", "Three", 0, 0)]
+        [InlineData("", "Wrong", 0, 0)]
+        public async Task CacheSearchPackageFilterTest(string commandTemplate, string commandPackage, int packMatchCount, int templateMatchCount)
+        {
+            TemplateDiscoveryMetadata mockTemplateDiscoveryMetadata = SetupDiscoveryMetadata(false);
+            MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
+            EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
+
+            MockNewCommandInput commandInput = new MockNewCommandInput(commandTemplate).WithCommandOption("--package", commandPackage);
+
+            TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
+            SearchResults searchResults = await searchCoordinator.SearchAsync();
+
+            Assert.True(searchResults.AnySources);
+            if (packMatchCount == 0)
+            {
+                Assert.Equal(0, searchResults.MatchesBySource.Count);
+            }
+            else
+            {
+                Assert.Equal(1, searchResults.MatchesBySource.Count);
+                Assert.Equal(packMatchCount, searchResults.MatchesBySource[0].PacksWithMatches.Count);
+                Assert.Equal(templateMatchCount, searchResults.MatchesBySource[0].PacksWithMatches[_packThreeInfo].TemplateMatches.Count);
+            }
+        }
+
         [Fact(DisplayName = nameof(CacheSearchLanguageMismatchFilterTest))]
         public async Task CacheSearchLanguageMismatchFilterTest()
         {
@@ -144,12 +197,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
             MockCliNuGetMetadataSearchSource.SetupMockData(mockTemplateDiscoveryMetadata);
             EngineEnvironmentSettings.SettingsLoader.Components.Register(typeof(MockCliNuGetMetadataSearchSource));
 
-            Dictionary<string, string> rawCommandInputs = new Dictionary<string, string>();
-            MockNewCommandInput commandInput = new MockNewCommandInput(rawCommandInputs)
-            {
-                TemplateName = "bar",
-                Language = "VB"
-            };
+            MockNewCommandInput commandInput = new MockNewCommandInput("bar", "VB");
 
             TemplateSearchCoordinator searchCoordinator = CliTemplateSearchCoordinatorFactory.CreateCliTemplateSearchCoordinator(EngineEnvironmentSettings, commandInput, DefaultLanguage);
             SearchResults searchResults = await searchCoordinator.SearchAsync();
@@ -162,75 +210,29 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests
         private static readonly PackInfo _packTwoInfo = new PackInfo("PackTwo", "1.6.0");
         private static readonly PackInfo _packThreeInfo = new PackInfo("PackThree", "2.1");
 
-        private static readonly ITemplateInfo _fooOneTemplate = new TemplateInfo()
-        {
-            Identity = "Mock.Foo.1",
-            GroupIdentity = "Mock.Foo",
-            Description = "Mock Foo template one",
-            Name = "MockFooTemplateOne",
-            ShortName = "foo1",
-            Tags = new Dictionary<string, ICacheTag>(StringComparer.Ordinal)
-            {
-                {
-                    "Framework", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "netcoreapp3.0", "netcoreapp3.1" })
-                },
-                {
-                    "language", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "C#" })
-                }
-            },
-            CacheParameters = new Dictionary<string, ICacheParameter>()
-        };
+        private static readonly ITemplateInfo _fooOneTemplate =
+            new MockTemplateInfo("foo1", name: "MockFooTemplateOne", identity: "Mock.Foo.1", groupIdentity: "Mock.Foo", author: "TestAuthor")
+                .WithDescription("Mock Foo template one")
+                .WithTag("Framework", "netcoreapp3.0", "netcoreapp3.1")
+                .WithTag("language", "C#")
+                .WithTag("type", "project");
 
-        private static readonly ITemplateInfo _fooTwoTemplate = new TemplateInfo()
-        {
-            Identity = "Mock.Foo.2",
-            GroupIdentity = "Mock.Foo",
-            Description = "Mock Foo template two",
-            Name = "MockFooTemplateTwo",
-            ShortName = "foo2",
-            Tags = new Dictionary<string, ICacheTag>(StringComparer.Ordinal)
-            {
-                {
-                    "Framework", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "netcoreapp2.0", "netcoreapp2.1", "netcoreapp3.1" })
-                },
-                {
-                    "language", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "C#" })
-                }
-            },
-            CacheParameters = new Dictionary<string, ICacheParameter>()
-        };
+        private static readonly ITemplateInfo _fooTwoTemplate =
+            new MockTemplateInfo("foo2", name: "MockFooTemplateTwo", identity: "Mock.Foo.2", groupIdentity: "Mock.Foo")
+                .WithDescription("Mock Foo template two")
+                .WithTag("Framework", "netcoreapp2.0", "netcoreapp2.1", "netcoreapp3.1")
+                .WithTag("language", "C#");
 
-        private static readonly ITemplateInfo _barCSharpTemplate = new MockTemplateInfo()
-        {
-            Identity = "Mock.Bar.1.Csharp",
-            GroupIdentity = "Mock.Bar",
-            Description = "Mock Bar CSharp template",
-            Name = "MockBarCsharpTemplate",
-            ShortName = "barC",
-            Tags = new Dictionary<string, ICacheTag>(StringComparer.Ordinal)
-            {
-                {
-                    "language", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "C#" })
-                }
-            },
-            CacheParameters = new Dictionary<string, ICacheParameter>()
-        };
 
-        private static readonly ITemplateInfo _barFSharpTemplate = new MockTemplateInfo()
-        {
-            Identity = "Mock.Bar.1.FSharp",
-            GroupIdentity = "Mock.Bar",
-            Description = "Mock Bar FSharp template",
-            Name = "MockBarFSharpTemplate",
-            ShortName = "barF",
-            Tags = new Dictionary<string, ICacheTag>(StringComparer.Ordinal)
-            {
-                {
-                    "language", ResolutionTestHelper.CreateTestCacheTag(new List<string>() { "F#" })
-                }
-            },
-            CacheParameters = new Dictionary<string, ICacheParameter>()
-        };
+        private static readonly ITemplateInfo _barCSharpTemplate =
+            new MockTemplateInfo("barC", name: "MockBarCsharpTemplate", identity: "Mock.Bar.1.Csharp", groupIdentity: "Mock.Bar")
+                .WithDescription("Mock Bar CSharp template")
+                .WithTag("language", "C#");
+
+        private static readonly ITemplateInfo _barFSharpTemplate =
+            new MockTemplateInfo("barF", name: "MockBarFSharpTemplate", identity: "Mock.Bar.1.FSharp", groupIdentity: "Mock.Bar")
+                .WithDescription("Mock Bar FSharp template")
+                .WithTag("language", "F#");
 
         private static TemplateDiscoveryMetadata SetupDiscoveryMetadata(bool includehostData = false)
         {
